requirements.txt
pyTelegramBotAPI==3.7.3
Pillow==7.2.0
requests==2.24.0